+++ apps/ledger-live-desktop/src/newArch/features/ModularDrawer/screens/AccountsWarning/types.ts
@@ -16,5 +16,6 @@ export interface ActionButtonsProps {
 export interface AccountsWarningProps {
   warningReason: WarningReason;
   currency: CryptoCurrency;
+  navigateToFundAccount: (account: Account) => void;
   emptyAccount?: Account;
 }